requirements.txt
# MindGuard AI - Requirements
# Production dependencies for the mental health ML pipeline

# Core Data Processing
pandas==2.1.3
numpy==1.24.3

# Reddit API
praw==7.7.0

# Configuration Management
python-dotenv==1.0.0

# Security & Encryption
cryptography==41.0.7

# Type Checking
mypy==1.7.1
types-requests==2.31.0.10

# Logging & Monitoring
python-json-logger==2.0.7
wandb==0.16.3              # W&B for experiment tracking and monitoring

# ML Model Inference (RoBERTa)
torch==2.1.2
transformers==4.35.2
scikit-learn==1.3.2

# Visualization & Dashboard
streamlit==1.29.0          # Interactive dashboard
plotly==5.18.0             # Advanced plotting
streamlit-plotly-events==0.0.6  # Plotly integration with Streamlit

# Async Support (optional, for future scaling)
asyncpraw==7.7.0
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
pylint==3.0.3
isort==5.13.2

# Database (optional, for production deployment)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# API & Web Framework
fastapi==0.109.0
uvicorn==0.27.0
